refactor: BaseException 추상클래스 제거
- 바로 CustomException 클래스로 생성하도록 변경
- exception 패키지를 enum 패키지로 이름 변경
