--- src/main/java/imhong/dowith/challenge/exception/ChallengeException.java
@@ -1,11 +1,0 @@
-package imhong.dowith.challenge.exception;
-
-import imhong.dowith.common.BaseException;
-import imhong.dowith.common.BaseExceptionType;
-
-public class ChallengeException extends BaseException {
-
-    public ChallengeException(BaseExceptionType baseExceptionType) {
-        super(baseExceptionType);
-    }
-}
